requirements.txt
aadict==0.2.3
absl-py==1.0.0
aiobotocore==1.2.2
aiohttp==3.7.4.post0
aioitertools==0.7.1
alabaster==0.7.12
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arviz==0.11.4
asset==0.6.13
async-generator==1.10
async-timeout==3.0.1
attrs==19.3.0
Automat==20.2.0
autopep8==1.4.4
Babel==2.8.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.9.1
bleach==4.1.0
boto3==1.14.39
botocore==1.17.39
cachetools==4.2.1
certifi==2019.3.9
cffi==1.14.0
cftime==1.5.1.1
chardet==3.0.4
colorama==0.4.4
commonmark==0.9.1
configparser==5.0.2
constantly==15.1.0
crayons==0.4.0
cryptography==3.4.7
cssselect==1.1.0
cssutils==2.2.0
cycler==0.10.0
Cython==0.29.22
dataclasses==0.8
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.3
Django==2.2.10
django-allauth==0.39.1
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
Dmail==1.2.6
dnslib==0.9.14
dnspython==1.16.0
docopt==0.6.2
docutils==0.15.2
dulwich==0.20.21
EasyProcess==0.3
entrypoints==0.3
fabric==2.5.0
flake8==3.8.4
flake8-isort==4.0.0
flatbuffers==2.0
fsspec==0.8.7
future==0.18.2
futures==3.1.1
gitdb==4.0.5
GitPython==3.1.3
globre==0.1.5
helium==3.0.4
html2text==2020.1.16
html5lib==1.1
hyperlink==19.0.0
idna==2.8
idna-ssl==1.1.0
imagesize==1.2.0
importlib-metadata==1.6.1
incremental==17.5.0
invoke==1.4.1
ipdb==0.13.2
ipykernel==5.5.6
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.8.0
jax==0.2.17
jaxlib==0.1.69
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==1.1.0
jsonpickle==1.4.1
jsonschema==3.2.0
jupyter-client==7.1.1
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.2.0
localstack==0.12.7
localstack-client==1.14
localstack-ext==0.12.6.1
lxml==4.5.1
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
Mini-AMF==0.9.1
mistune==0.8.4
mmh3==3.0.0
more-itertools==8.3.0
multidict==5.1.0
multiprocess==0.70.11.1
natsort==8.1.0
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.4
netCDF4==1.5.8
notebook==6.4.7
numpy==1.19.5
numpyro==0.7.2
oauthlib==3.0.1
opt-einsum==3.3.0
packaging==20.4
pandas==1.0.5
pandocfilters==1.5.0
paramiko==2.7.1
parso==0.7.0
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.2.0
pipreqs==0.4.10
pluggy==0.13.1
plyvel==1.3.0
premailer==3.7.0
probableparsing==0.0.1
prometheus-client==0.12.0
prompt-toolkit==3.0.5
protobuf==3.12.2
psutil==5.8.0
ptyprocess==0.6.0
publicsuffix==1.1.1
py==1.8.1
pyaes==1.6.1
pyarrow==3.0.0
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.6.1
PyHamcrest==2.0.2
pyhocon==0.3.55
PyNaCl==1.4.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==5.4.3
pytest-subtests==0.3.1
python-crfsuite==0.9.7
python-dateutil==2.8.1
python-decouple==3.1
python3-openid==3.1.0
pytils==0.3
pytz==2018.5
PyVirtualDisplay==1.3.2
PyYAML==5.4.1
pyzmq==22.3.0
recommonmark==0.6.0
requests==2.21.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
rinoh-typeface-dejavuserif==0.1.1
rinoh-typeface-texgyrecursor==0.1.1
rinoh-typeface-texgyreheros==0.1.1
rinoh-typeface-texgyrepagella==0.1.1
rinohtype==0.4.2
s3fs==0.5.1
s3transfer==0.3.3
scikit-learn==0.24.2
scipy==1.5.4
seaborn==0.11.2
selenium==3.141.0
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
smmap==3.0.4
snowballstemmer==2.0.0
soupsieve==2.0.1
Sphinx==3.2.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-confluencebuilder==1.2.0
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.2.4
stripe==2.27.0
tabulate==0.8.9
tblib==1.7.0
tensorboard-logger==0.1.0
tensorboardX==2.0
terminado==0.12.1
testfixtures==6.17.1
testpath==0.5.0
threadpoolctl==3.1.0
tld==0.12.5
tools==0.1.9
torch==1.2.0+cpu
torchvision==0.4.0+cpu
tornado==6.1
tqdm==4.62.3
traitlets==4.3.3
Twisted==20.3.0
typing-extensions==3.7.4.3
urllib3==1.24.2
usaddress==0.5.10
wcwidth==0.2.4
webdriver-manager==3.3.0
webencodings==0.5.1
widgetsnbextension==3.5.2
wrapt==1.12.1
xarray==0.16.2
xvfbwrapper==0.2.9
yarg==0.1.9
yarl==1.6.3
zipp==3.1.0
zope.interface==5.1.0
